requirements.txt
chainlit
pandas
python-docx
pymupdf
python-pptx
beautifulsoup4
lxml
langchain-google-genai
langchain-core
langchain-community